// src/app/constants/marker-animation.enum.ts
export enum MarkerAnimationEnum {
    OPEN = '0px -5px 10px 10px rgba(0,255,0,0.5) inset, 0px -5px 10px 5px rgba(0,255,0,0.7)',
    CLOSED = '0px -5px 10px 10px rgba(255,0,0,0.5) inset, 0px -5px 10px 5px rgba(255,0,0,0.7)',
    WARNING = '0px -5px 10px 10px rgba(255,255,0,0.5) inset, 0px -5px 10px 5px rgba(255,255,0,0.7)',
    NONE = '',
}