// src/app/constants/marker-colour.enum.ts
export enum MarkerColourEnum {
    STATION = '#f7acef',
    SET = '#bff4e9',
    SECTOR = '#faddb0',
    UNIT_HYDRANT_LOW = '#c2f5b6',
    UNIT_HYDRANT_MEDIUM = '#71c15f',
    UNIT_HYDRANT_HIGTH = '#3c9128',
    UNIT_POND_LOW = '#c9b6f5',
    UNIT_POND_MEDIUM = '#7f5fc1',
    UNIT_POND_HIGTH = '#502891',
    ALARM = '#ff0000',
    WARNING = '#fffc00',

}